refactor(python): move python-specific logic to own module
