+++ exec/util.go
@@ -1,25 +1,10 @@
-package main
+package exec
 
 import (
-	"errors"
 	"os"
 	"os/exec"
 	"strings"
 )
-
-func VersionStringToStruct(version string) VersionTag {
-	splitVersion := strings.Split(version, ".")
-
-	return VersionTag{Major: splitVersion[0], Minor: splitVersion[1], Patch: splitVersion[2]}
-}
-
-func ValidateVersion(version string) error {
-	if splitVersion := strings.Split(version, "."); len(splitVersion) != 3 {
-		return errors.New("Invalid version string. Expected format 'a.b.c'.")
-	}
-
-	return nil
-}
 
 // RunCommand is a thin wrapper around running command-line calls
 // programmatically. It abstracts common configuration like routing
